# docs/swagger.yaml
basePath: /
definitions:
  requests.CreateProductRequest:
    properties:
      content:
        example: Echo is nice!
        type: string
      price:
        example: 40000
        type: integer
      title:
        example: Echo
        type: string
    required:
    - price
    - title
    type: object
  requests.LoginRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      password:
        example: password123456
        type: string
    required:
    - email
    - password
    type: object
  requests.RefreshRequest:
    properties:
      token:
        example: refresh_token
        type: string
    required:
    - token
    type: object
  requests.RegisterRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: password123456
        type: string
    required:
    - email
    - name
    - password
    type: object
  requests.UpdateProductRequest:
    properties:
      content:
        example: Echo is nice!
        type: string
      price:
        example: 40000
        type: integer
      title:
        example: Echo
        type: string
    required:
    - price
    - title
    type: object
  responses.Data:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  responses.Error:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
  responses.LoginResponse:
    properties:
      accessToken:
        type: string
      exp:
        type: integer
      refreshToken:
        type: string
    type: object
  responses.ProductResponse:
    properties:
      desc:
        example: Echo is nice and delicious
        type: string
      id:
        example: 1
        type: integer
      image_url:
        type: string
      price:
        example: 100000
        type: integer
      title:
        example: Echo
        type: string
      username:
        example: John Doe
        type: string
    type: object
info:
  contact:
    email: haikalvidya@gmail.com
    name: Vidya Haikal
  description: This is a simple version of e-commerce using Echo app.
  title: Go Ecomrce
  version: "1.0"
paths:
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Perform user login
      operationId: user-login
      parameters:
      - description: User's credentials
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Authenticate a user
      tags:
      - User Actions
  /api/v1/products:
    get:
      description: Get the list of all products
      operationId: products-get
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.ProductResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get products
      tags:
      - Products Actions
    post:
      consumes:
      - application/json
      description: Create product
      operationId: products-create
      parameters:
      - description: Product title, description, price, and image
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
      security:
      - ApiKeyAuth: []
      summary: Create product
      tags:
      - Products Actions
  /api/v1/products/{id}:
    delete:
      description: Delete product
      operationId: products-delete
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/responses.Data'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete product
      tags:
      - Products Actions
    get:
      description: Get the products by id
      operationId: products-get-one
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.ProductResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get product
      tags:
      - Products Actions
    put:
      consumes:
      - application/json
      description: Update product
      operationId: products-update
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product title, description, price, and image
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
      security:
      - ApiKeyAuth: []
      summary: Update product
      tags:
      - Products Actions
  /api/v1/products/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Upload Image product
      operationId: products-upload-image
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: this is a test file
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
      security:
      - ApiKeyAuth: []
      summary: Upload Image product
      tags:
      - Products Actions
  /api/v1/refresh:
    post:
      consumes:
      - application/json
      description: Perform refresh access token
      operationId: user-refresh
      parameters:
      - description: Refresh token
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Refresh access token
      tags:
      - User Actions
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: New user registration
      operationId: user-register
      parameters:
      - description: User's email, user's password
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Register
      tags:
      - User Actions
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"